# docs/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Top Organic API
  version: 1.0.0
  description: >
    Versioned API for Top Organic Enterprise Hybrid Restaurant ERP.
    Covers authentication, session, device, offline-login, RBAC, audit (Identity module)
    and catalog, shifts/drawers/tables, orders, customers, gift cards, billing,
    KDS, printing, and offline synchronization (Sales/POS module).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /pos/shifts:
    post:
      tags: [POS]
      operationId: openShift
      summary: Open POS shift (pos.device required)
      responses:
        '201': {$ref: '#/components/responses/Shift'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /pos/shifts/{shift}:
    parameters:
      - {$ref: '#/components/parameters/ShiftId'}
    get:
      tags: [POS]
      operationId: showShift
      summary: Show POS shift
      responses:
        '200': {$ref: '#/components/responses/Shift'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
  /pos/shifts/{shift}/close:
    post:
      tags: [POS]
      operationId: closeShift
      summary: Close POS shift (pos.device required)
      parameters:
        - {$ref: '#/components/parameters/ShiftId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpectedVersionRequest'}
      responses:
        '200': {$ref: '#/components/responses/Shift'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /pos/drawers/sessions:
    post:
      tags: [POS]
      operationId: openDrawer
      summary: Open cash drawer session (pos.device required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shift_id, drawer_id, currency, opening_amount]
              properties:
                shift_id: {$ref: '#/components/schemas/Ulid'}
                drawer_id: {$ref: '#/components/schemas/Ulid'}
                currency: {type: string, minLength: 3, maxLength: 3}
                opening_amount: {$ref: '#/components/schemas/MoneyAmount'}
      responses:
        '201': {$ref: '#/components/responses/DrawerSession'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /pos/drawers/sessions/{session}/close:
    post:
      tags: [POS]
      operationId: closeDrawer
      summary: Close cash drawer session (pos.device required)
      parameters:
        - {$ref: '#/components/parameters/SessionId'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [counted_amount, expected_version]
              properties:
                counted_amount: {$ref: '#/components/schemas/MoneyAmount'}
                expected_version: {$ref: '#/components/schemas/LockVersion'}
      responses:
        '200': {$ref: '#/components/responses/DrawerSession'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /pos/drawers/sessions/{session}/movements:
    post:
      tags: [POS]
      operationId: recordCashMovement
      summary: Record cash movement (pos.device required)
      parameters:
        - {$ref: '#/components/parameters/SessionId'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, amount, currency, client_operation_id]
              properties:
                type: {type: string, enum: [cash_in, cash_out, sale, refund, adjustment]}
                amount: {type: integer, description: Non-zero signed integer minor units.}
                currency: {type: string, minLength: 3, maxLength: 3}
                client_operation_id: {$ref: '#/components/schemas/ClientOperationId'}
                reason: {type: string, maxLength: 500, nullable: true}
      responses:
        '201': {$ref: '#/components/responses/GenericData'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /pos/cash-movements/{movement}/reverse:
    post:
      tags: [POS]
      operationId: reverseCashMovement
      summary: Reverse a cash movement (pos.device required)
      parameters:
        - name: movement
          in: path
          required: true
          schema: {$ref: '#/components/schemas/Ulid'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, client_operation_id]
              properties:
                reason: {type: string, maxLength: 500}
                client_operation_id: {$ref: '#/components/schemas/ClientOperationId'}
      responses:
        '201': {$ref: '#/components/responses/GenericData'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
...
  /pos/table-sessions:
    post:
      tags: [POS]
      operationId: openTableSession
      summary: Open table session (pos.device required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [table_id, guest_count]
              properties:
                table_id: {$ref: '#/components/schemas/Ulid'}
                guest_count: {type: integer, minimum: 1, maximum: 1000}
      responses:
        '201': {$ref: '#/components/responses/TableSession'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /pos/table-sessions/{session}/close:
    post:
      tags: [POS]
      operationId: closeTableSession
      summary: Close table session (pos.device required)
      parameters:
        - {$ref: '#/components/parameters/SessionId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpectedVersionRequest'}
      responses:
        '200': {$ref: '#/components/responses/TableSession'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}

  # ─── Orders ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived API access JWT only. Opaque refresh tokens and offline JWTs are not accepted here.
  parameters:
...
    SessionId:
      name: session
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Ulid'}
...
  responses:
    Error:
      description: Stable error envelope.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$'
      example: 01J00000000000000000000000
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, example: UNAUTHENTICATED}
            message: {type: string}
            details: {type: object, additionalProperties: true}
...